+++ docs/CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Change password flow.
+- Version info section to profile view.
 
 ## [1.1.0] - 2022-04-19